requirements.txt
pandas
openpyxl
snowflake-connector-python
snowflake-snowpark-python
